Move hint_utils to builtin_hint_processor
