+++ src/hint_processor/builtin_hint_processor/blake2s_utils.rs
@@ -5,14 +5,14 @@ use num_traits::ToPrimitive;
 use super::blake2s_hash::blake2s_compress;
 use crate::bigint;
 use crate::hint_processor::builtin_hint_processor::blake2s_hash::IV;
-use crate::hint_processor::hint_processor_definition::HintReference;
-use crate::hint_processor::hint_utils::{
+use crate::hint_processor::builtin_hint_processor::hint_utils::{
     bigint_to_u32, get_ptr_from_var_name, get_relocatable_from_var_name,
 };
+use crate::hint_processor::hint_processor_definition::HintReference;
+use crate::hint_processor::proxies::memory_proxy::MemoryProxy;
+use crate::hint_processor::proxies::vm_proxy::VMProxy;
 use crate::serde::deserialize_program::ApTracking;
 use crate::types::relocatable::Relocatable;
-use crate::vm::vm_core::VMProxy;
-use crate::vm::vm_memory::memory::MemoryProxy;
 use crate::{
     types::relocatable::MaybeRelocatable,
     vm::{
@@ -252,11 +252,11 @@ pub fn blake2s_add_uint256_bigend(
 mod tests {
     use super::*;
     use crate::any_box;
-    use crate::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::get_vm_proxy;
     use crate::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::BuiltinHintProcessor;
     use crate::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::HintProcessorData;
+    use crate::hint_processor::proxies::exec_scopes_proxy::get_exec_scopes_proxy;
+    use crate::hint_processor::proxies::vm_proxy::get_vm_proxy;
     use crate::relocatable;
-    use crate::types::exec_scope::get_exec_scopes_proxy;
     use crate::types::exec_scope::ExecutionScopes;
     use crate::utils::test_utils::*;
     use crate::vm::vm_core::VirtualMachine;
