+++ src/hint_processor/builtin_hint_processor/cairo_keccak/keccak_hints.rs
@@ -1,12 +1,11 @@
-use crate::hint_processor::hint_utils::get_integer_from_var_name;
-use crate::hint_processor::hint_utils::get_ptr_from_var_name;
-use crate::hint_processor::hint_utils::insert_value_into_ap;
+use crate::hint_processor::builtin_hint_processor::hint_utils::get_integer_from_var_name;
+use crate::hint_processor::builtin_hint_processor::hint_utils::get_ptr_from_var_name;
+use crate::hint_processor::builtin_hint_processor::hint_utils::insert_value_into_ap;
+use crate::hint_processor::proxies::vm_proxy::VMProxy;
 use crate::{
-    bigint,
-    hint_processor::hint_processor_definition::HintReference,
-    serde::deserialize_program::ApTracking,
-    types::relocatable::MaybeRelocatable,
-    vm::{errors::vm_errors::VirtualMachineError, vm_core::VMProxy},
+    bigint, hint_processor::hint_processor_definition::HintReference,
+    serde::deserialize_program::ApTracking, types::relocatable::MaybeRelocatable,
+    vm::errors::vm_errors::VirtualMachineError,
 };
 use lazy_static::lazy_static;
 use num_bigint::BigInt;
@@ -247,12 +246,12 @@ fn u64_array_to_bigint_vec(array: &[u64; KECCAK_STATE_SIZE_FELTS]) -> Vec<BigInt
 mod tests {
     use super::*;
     use crate::any_box;
-    use crate::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::get_vm_proxy;
     use crate::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::BuiltinHintProcessor;
     use crate::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::HintProcessorData;
     use crate::hint_processor::hint_processor_definition::HintProcessor;
     use crate::hint_processor::hint_processor_definition::HintReference;
-    use crate::types::exec_scope::get_exec_scopes_proxy;
+    use crate::hint_processor::proxies::exec_scopes_proxy::get_exec_scopes_proxy;
+    use crate::hint_processor::proxies::vm_proxy::get_vm_proxy;
     use crate::types::exec_scope::ExecutionScopes;
     use crate::utils::test_utils::*;
     use crate::vm::errors::memory_errors::MemoryError;
