+++ src/hint_processor/builtin_hint_processor/find_element_hint.rs
@@ -1,13 +1,15 @@
-use crate::hint_processor::hint_processor_definition::HintReference;
-use crate::hint_processor::hint_utils::{
+use crate::hint_processor::builtin_hint_processor::hint_utils::{
     bigint_to_usize, get_ptr_from_var_name, get_relocatable_from_var_name,
     insert_value_from_var_name,
 };
+use crate::hint_processor::hint_processor_definition::HintReference;
+use crate::hint_processor::proxies::exec_scopes_proxy::ExecutionScopesProxy;
+use crate::hint_processor::proxies::vm_proxy::VMProxy;
 use crate::serde::deserialize_program::ApTracking;
-use crate::types::exec_scope::ExecutionScopesProxy;
 use crate::vm::errors::vm_errors::VirtualMachineError;
-use crate::vm::vm_core::VMProxy;
-use crate::{bigint, hint_processor::hint_utils::get_integer_from_var_name};
+use crate::{
+    bigint, hint_processor::builtin_hint_processor::hint_utils::get_integer_from_var_name,
+};
 use num_bigint::BigInt;
 use num_traits::{Signed, ToPrimitive};
 use std::collections::HashMap;
@@ -149,12 +151,12 @@ pub fn search_sorted_lower(
 mod tests {
     use super::*;
     use crate::any_box;
-    use crate::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::get_vm_proxy;
     use crate::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::BuiltinHintProcessor;
     use crate::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::HintProcessorData;
     use crate::hint_processor::builtin_hint_processor::hint_code;
     use crate::hint_processor::hint_processor_definition::HintProcessor;
-    use crate::types::exec_scope::get_exec_scopes_proxy;
+    use crate::hint_processor::proxies::exec_scopes_proxy::get_exec_scopes_proxy;
+    use crate::hint_processor::proxies::vm_proxy::get_vm_proxy;
     use crate::types::exec_scope::ExecutionScopes;
     use crate::types::relocatable::MaybeRelocatable;
     use crate::utils::test_utils::vm;
