+++ src/hint_processor/builtin_hint_processor/uint256_utils.rs
@@ -1,17 +1,21 @@
 use crate::bigint;
-use crate::hint_processor::hint_utils::{get_integer_from_var_name, get_relocatable_from_var_name};
+use crate::hint_processor::builtin_hint_processor::hint_utils::{
+    get_integer_from_var_name, get_relocatable_from_var_name,
+};
+use crate::hint_processor::proxies::vm_proxy::VMProxy;
 use crate::math_utils::isqrt;
 use crate::serde::deserialize_program::ApTracking;
 use crate::vm::errors::vm_errors::VirtualMachineError;
-use crate::vm::vm_core::VMProxy;
 use num_bigint::BigInt;
 use num_integer::{div_rem, Integer};
 use num_traits::Signed;
 use std::collections::HashMap;
 use std::ops::{Shl, Shr};
 
+use crate::hint_processor::builtin_hint_processor::hint_utils::{
+    insert_value_from_var_name, insert_value_into_ap,
+};
 use crate::hint_processor::hint_processor_definition::HintReference;
-use crate::hint_processor::hint_utils::{insert_value_from_var_name, insert_value_into_ap};
 
 /*
 Implements hint:
@@ -216,11 +220,11 @@ mod tests {
     use super::*;
     use crate::any_box;
     use crate::bigint_str;
-    use crate::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::get_vm_proxy;
     use crate::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::BuiltinHintProcessor;
     use crate::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::HintProcessorData;
     use crate::hint_processor::hint_processor_definition::HintProcessor;
-    use crate::types::exec_scope::get_exec_scopes_proxy;
+    use crate::hint_processor::proxies::exec_scopes_proxy::get_exec_scopes_proxy;
+    use crate::hint_processor::proxies::vm_proxy::get_vm_proxy;
     use crate::types::exec_scope::ExecutionScopes;
     use crate::types::relocatable::MaybeRelocatable;
     use crate::utils::test_utils::*;
