+++ src/hint_processor/hint_processor_definition.rs
@@ -1,22 +1,11 @@
+use crate::hint_processor::proxies::exec_scopes_proxy::ExecutionScopesProxy;
+use crate::hint_processor::proxies::vm_proxy::VMProxy;
 use crate::serde::deserialize_program::ApTracking;
-use crate::types::exec_scope::ExecutionScopesProxy;
 use crate::types::instruction::Register;
 use crate::vm::errors::vm_errors::VirtualMachineError;
-use crate::vm::vm_core::VMProxy;
 use num_bigint::BigInt;
 use std::any::Any;
 use std::collections::HashMap;
-
-#[derive(Debug, PartialEq, Clone)]
-pub struct HintReference {
-    pub register: Register,
-    pub offset1: i32,
-    pub offset2: i32,
-    pub dereference: bool,
-    pub inner_dereference: bool,
-    pub ap_tracking_data: Option<ApTracking>,
-    pub immediate: Option<BigInt>,
-}
 
 pub trait HintProcessor {
     //Executes the hint which's data is provided by a dynamic structure previously created by compile_hint
@@ -45,4 +34,41 @@ pub trait HintProcessor {
         //List of all references (key corresponds to element of the previous dictionary)
         references: &HashMap<usize, HintReference>,
     ) -> Result<Box<dyn Any>, VirtualMachineError>;
+}
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct HintReference {
+    pub register: Register,
+    pub offset1: i32,
+    pub offset2: i32,
+    pub dereference: bool,
+    pub inner_dereference: bool,
+    pub ap_tracking_data: Option<ApTracking>,
+    pub immediate: Option<BigInt>,
+}
+
+impl HintReference {
+    pub fn new_simple(offset1: i32) -> Self {
+        HintReference {
+            register: Register::FP,
+            offset1,
+            offset2: 0,
+            inner_dereference: false,
+            ap_tracking_data: None,
+            immediate: None,
+            dereference: true,
+        }
+    }
+
+    pub fn new(offset1: i32, offset2: i32, inner_dereference: bool, dereference: bool) -> Self {
+        HintReference {
+            register: Register::FP,
+            offset1,
+            offset2,
+            inner_dereference,
+            ap_tracking_data: None,
+            immediate: None,
+            dereference,
+        }
+    }
 }
