+++ src/hint_processor/mod.rs
@@ -1,3 +1,3 @@
 pub mod builtin_hint_processor;
 pub mod hint_processor_definition;
-pub mod hint_utils;
+pub mod proxies;
